requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
argparse==1.2.1
gunicorn==19.1.1
itsdangerous==0.24
six==1.8.0
stevedore==1.0.0
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
python-twitter>=3.3
cfscrape>=1.9.5



beautifulsoup4>=4.4.1
Pillow>=3.3.0
PyYAML>=3.11
cssselect>=0.9.2
lxml>=3.6.0
nltk>=3.2.1
requests>=2.10.0
feedparser>=5.2.1
tldextract>=2.0.1
feedfinder2>=0.0.4
jieba3k>=0.35.1
python-dateutil>=2.5.3


twitterscraper>=0.7.1